+++ libnavui-dropin/src/main/java/com/mapbox/navigation/dropin/tripsession/TripSessionComponent.kt
@@ -3,7 +3,9 @@ package com.mapbox.navigation.dropin.tripsession
 import android.annotation.SuppressLint
 import androidx.lifecycle.Lifecycle
 import androidx.lifecycle.repeatOnLifecycle
+import com.mapbox.navigation.base.ExperimentalPreviewMapboxNavigationAPI
 import com.mapbox.navigation.core.MapboxNavigation
+import com.mapbox.navigation.core.replay.route.ReplayRouteSession
 import com.mapbox.navigation.core.trip.session.TripSessionState
 import com.mapbox.navigation.ui.app.internal.SharedApp.tripSessionTransaction
 import com.mapbox.navigation.ui.app.internal.Store
@@ -16,13 +18,14 @@ import kotlinx.coroutines.flow.combine
 import kotlinx.coroutines.flow.distinctUntilChanged
 import kotlinx.coroutines.launch
 
+@OptIn(ExperimentalPreviewMapboxNavigationAPI::class)
 @SuppressLint("MissingPermission")
 internal class TripSessionComponent(
     private val lifecycle: Lifecycle,
     private val store: Store
 ) : UIComponent() {
 
-    private var replayRouteTripSession: ReplayRouteTripSession? = null
+    private var replayRouteTripSession: ReplayRouteSession? = null
 
     /**
      * Signals that the [mapboxNavigation] instance is ready for use.
@@ -57,7 +60,7 @@ internal class TripSessionComponent(
     override fun onDetached(mapboxNavigation: MapboxNavigation) {
         super.onDetached(mapboxNavigation)
         tripSessionTransaction {
-            replayRouteTripSession?.stop(mapboxNavigation)
+            replayRouteTripSession?.onDetached(mapboxNavigation)
             replayRouteTripSession = null
         }
     }
@@ -74,15 +77,15 @@ internal class TripSessionComponent(
     }.distinctUntilChanged()
 
     private fun startTripSession(mapboxNavigation: MapboxNavigation) {
-        replayRouteTripSession?.stop(mapboxNavigation)
+        replayRouteTripSession?.onDetached(mapboxNavigation)
         replayRouteTripSession = null
         mapboxNavigation.ensureTripSessionStarted()
     }
 
     private fun startReplayTripSession(mapboxNavigation: MapboxNavigation) {
-        replayRouteTripSession?.stop(mapboxNavigation)
-        replayRouteTripSession = ReplayRouteTripSession()
-        replayRouteTripSession?.start(mapboxNavigation)
+        replayRouteTripSession?.onDetached(mapboxNavigation)
+        replayRouteTripSession = ReplayRouteSession()
+        replayRouteTripSession?.onAttached(mapboxNavigation)
     }
 
     private fun MapboxNavigation.ensureTripSessionStarted() {